+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added missing match code index
 - altered maps json column to LONGTEXT to fit big maps
 - added maps listing
+- added map preview as png
 
 ### 0.2.0-beta
 - updated database schema